Made better spaced name handling and fixed bugs
Fixed issues with spaces in names being treated as though they aren't there resulting in the name On Hyex (with one space) treated the same of On  Hyex or On  Hy ex (two space for the first space). Also fixed errors with using playerIDs and fixed issues with having no message attached. Also fixed errors resulting from the spaces being ignored resulting in the name being partially displayed in message or some of the message being missing if using the incorrect number of spaces in the name.